+++ modules/snooker/src/test/ts/browser/ResizeTest.ts
@@ -1,20 +1,27 @@
-import { assert, UnitTest } from '@ephox/bedrock-client';
+import { after, before, describe, it } from '@ephox/bedrock-client';
 import { Arr } from '@ephox/katamari';
 import { Css, Insert, Remove, SugarBody, SugarElement } from '@ephox/sugar';
+import { assert } from 'chai';
 
 import * as ResizeBehaviour from 'ephox/snooker/api/ResizeBehaviour';
 import { TableSize } from 'ephox/snooker/api/TableSize';
 import { Warehouse } from 'ephox/snooker/api/Warehouse';
 import * as Deltas from 'ephox/snooker/calc/Deltas';
 
-UnitTest.test('ResizeTest', () => {
+describe('ResizeTest', () => {
   const resizing = ResizeBehaviour.preserveTable();
-
   const boundBox = '<div style="width: 800px; height: 600px; display: block;"></div>';
   const box = SugarElement.fromHtml<HTMLDivElement>(boundBox);
-  Insert.append(SugarBody.body(), box);
 
-  const percentTablePercentCellsTest = () => {
+  before(() => {
+    Insert.append(SugarBody.body(), box);
+  });
+
+  after(() => {
+    Remove.remove(box);
+  });
+
+  it('should be able to resize a percent table with percent cells', () => {
     const delta = 200;
 
     const table = SugarElement.fromHtml<HTMLTableElement>(`<table style="border-collapse: collapse; width: 100%;">
@@ -39,31 +46,31 @@ UnitTest.test('ResizeTest', () => {
     const tableSize = TableSize.getTableSize(table);
 
     // 100% width table
-    assert.eq(100, tableSize.width());
+    assert.equal(100, tableSize.width());
 
     // 25% width delta
     const step = tableSize.getCellDelta(delta);
-    assert.eq(25, step);
+    assert.equal(25, step);
 
     const warehouse = Warehouse.fromTable(table);
     const widths = tableSize.getWidths(warehouse, tableSize);
 
     // [50%, 50%] existing widths.
-    assert.eq([ 50, 50 ], widths);
+    assert.deepEqual([ 50, 50 ], widths);
 
     const deltas = Deltas.determine(widths, 0, step, tableSize, resizing);
 
     // [25%, -25%] deltas.
-    assert.eq([ 25, -25 ], deltas);
+    assert.deepEqual([ 25, -25 ], deltas);
 
     // Set new width
     tableSize.adjustTableWidth(step);
-    assert.eq(Css.getRaw(table, 'width').getOrDie(), '125%');
+    assert.equal(Css.getRaw(table, 'width').getOrDie(), '125%');
 
     Remove.remove(table);
-  };
+  });
 
-  const percentTablePixelCellsTest = () => {
+  it('should be able to resize a percent table with pixel cells', () => {
     const delta = 200;
 
     const table = SugarElement.fromHtml<HTMLTableElement>(`<table style="border-collapse: collapse; width: 100%;">
@@ -88,40 +95,36 @@ UnitTest.test('ResizeTest', () => {
     const tableSize = TableSize.getTableSize(table);
 
     // 100% width table
-    assert.eq(100, tableSize.width());
+    assert.equal(100, tableSize.width());
 
     // 25% width delta
     const step = tableSize.getCellDelta(delta);
-    assert.eq(25, step);
+    assert.equal(25, step);
 
     const warehouse = Warehouse.fromTable(table);
     const widths = tableSize.getWidths(warehouse, tableSize);
-
-    const expectedWidths = [ 50, 50 ];
-
-    const widthDiffs = Arr.map(expectedWidths, (x, i) => widths[i] - x);
 
     // percentage width of this table is 100% but phantom treats this as around 804 pixels when we're doing conversions
     // we have pixel width cells of 400px, so the actual widths of the cells in percentages
     // in order for us to pass this test, we ensure that the difference between what we wanted (50%)
     // and the actual (50.125% and 49.825% respectively) are within a tolerance of 1%
-    Arr.each(widthDiffs, (x) => {
-      assert.eq(true, x < 1 && x > -1);
+    Arr.each([ 50, 50 ], (expected, i) => {
+      assert.approximately(widths[i], expected, 1);
     });
 
     const deltas = Deltas.determine(widths, 0, step, tableSize, resizing);
 
     // [25%, -25%] deltas.
-    assert.eq([ 25, -25 ], deltas);
+    assert.deepEqual([ 25, -25 ], deltas);
 
     // Set new width
     tableSize.adjustTableWidth(step);
-    assert.eq(Css.getRaw(table, 'width').getOrDie(), '125%');
+    assert.equal(Css.getRaw(table, 'width').getOrDie(), '125%');
 
     Remove.remove(table);
-  };
+  });
 
-  const pixelTablePixelCellsTest = () => {
+  it('should be able to resize a pixel table with pixel cells', () => {
     const delta = 200;
 
     const table = SugarElement.fromHtml<HTMLTableElement>(`<table style="border-collapse: collapse; width: 800px;">
@@ -146,27 +149,27 @@ UnitTest.test('ResizeTest', () => {
     const tableSize = TableSize.getTableSize(table);
 
     // 100% width table
-    assert.eq(800, tableSize.width());
+    assert.equal(800, tableSize.width());
 
     // 25% width delta
     const step = tableSize.getCellDelta(delta);
-    assert.eq(200, step);
+    assert.equal(200, step);
 
     const warehouse = Warehouse.fromTable(table);
     const widths = tableSize.getWidths(warehouse, tableSize);
 
     // [50%, 50%] existing widths.
-    assert.eq([ 400, 400 ], widths);
+    assert.deepEqual([ 398, 398 ], widths);
 
     const deltas = Deltas.determine(widths, 0, step, tableSize, resizing);
 
     // [25%, -25%] deltas.
-    assert.eq([ 200, -200 ], deltas);
+    assert.deepEqual([ 200, -200 ], deltas);
 
     Remove.remove(table);
-  };
+  });
 
-  const pixelTablePercentCellsTest = () => {
+  it('should be able to resize a pixel table with percent cells', () => {
     const delta = 200;
 
     const table = SugarElement.fromHtml<HTMLTableElement>(`<table style="border-collapse: collapse; width: 800px;">
@@ -191,29 +194,64 @@ UnitTest.test('ResizeTest', () => {
     const tableSize = TableSize.getTableSize(table);
 
     // 100% width table
-    assert.eq(800, tableSize.width());
+    assert.equal(800, tableSize.width());
 
     // 25% width delta
     const step = tableSize.getCellDelta(delta);
-    assert.eq(200, step);
+    assert.equal(200, step);
 
     const warehouse = Warehouse.fromTable(table);
     const widths = tableSize.getWidths(warehouse, tableSize);
 
     // [50%, 50%] existing widths.
-    assert.eq([ 400, 400 ], widths);
+    assert.deepEqual([ 398, 398 ], widths);
 
     const deltas = Deltas.determine(widths, 0, step, tableSize, resizing);
 
     // [25%, -25%] deltas.
-    assert.eq([ 200, -200 ], deltas);
+    assert.deepEqual([ 200, -200 ], deltas);
 
     Remove.remove(table);
-  };
+  });
 
-  percentTablePercentCellsTest();
-  percentTablePixelCellsTest();
-  pixelTablePixelCellsTest();
-  pixelTablePercentCellsTest();
-  Remove.remove(box);
+  it('TINY-7731: should handle resizing a table where a cell overflows it\'s specified size', () => {
+    const delta = 200;
+
+    const table = SugarElement.fromHtml<HTMLTableElement>(`<table style="border-collapse: collapse; width: 800px;">
+    <tbody>
+    <tr>
+    <td style="width: 400px;">thisisareallylongsentencewithoutspacesthatcausescontenttooverflow</td>
+    <td style="width: 400px;">B</td>
+    </tr>
+    <tr>
+    <td style="width: 400px;">C</td>
+    <td style="width: 400px;">D</td>
+    </tr>
+    </tbody>
+    </table>`);
+
+    Insert.append(box, table);
+
+    const tableSize = TableSize.getTableSize(table);
+
+    // 100% width table
+    assert.equal(800, tableSize.width());
+
+    // 25% width delta
+    const step = tableSize.getCellDelta(delta);
+    assert.equal(200, step);
+
+    const warehouse = Warehouse.fromTable(table);
+    const widths = tableSize.getWidths(warehouse, tableSize);
+
+    assert.approximately(widths[0], 483, 1, 'First column width');
+    assert.approximately(widths[1], 313, 1, 'Second column width');
+
+    const deltas = Deltas.determine(widths, 0, step, tableSize, resizing);
+
+    // [25%, -25%] deltas.
+    assert.deepEqual([ 200, -200 ], deltas);
+
+    Remove.remove(table);
+  });
 });