+++ modules/sugar/src/main/ts/ephox/sugar/api/search/Traverse.ts
@@ -36,6 +36,10 @@ const parent = (element: SugarElement<Node>): Optional<SugarElement<Node & Paren
 const parentNode = (element: SugarElement<Node>): Optional<SugarElement<Node>> =>
   parent(element) as any;
 
+// Note: This requires an Element since IE 11 doesn't support using `parentElement` on a `Node`
+const parentElement = (element: SugarElement<Element>): Optional<SugarElement<HTMLElement>> =>
+  Optional.from(element.dom.parentElement).map(SugarElement.fromDom);
+
 const findIndex = (element: SugarElement<Node>): Optional<number> =>
   parent(element).bind((p) => {
     // TODO: Refactor out children so we can avoid the constant unwrapping
@@ -129,6 +133,7 @@ export {
   documentElement,
   parent,
   parentNode,
+  parentElement,
   findIndex,
   parents,
   siblings,